Fix for building against musl libc
This fix enables the package to be built against
[musl libc](https://www.musl-libc.org/) instead of glibc. Musl libc is
used in the Linux Alpine distro and since the
[official Node.js Docker repo](https://hub.docker.com/_/node/)
advertises its Alpine based images as "highly recommended
when final image size being as small as possible is desired", it turns
out that quite a few people run Node.js on Alpine via these images.

This change should not have any impact on systems that use the more
widespread glibc, as the updated imports should be available in all
libc variants.